move Workspace implementations to their own submodule 'tango.workspaces'
