+++ tango/workspaces/__init__.py
@@ -1,0 +1,6 @@
+"""
+Built-in :class:`~tango.workspace.Workspace` implementations.
+"""
+
+from .local_workspace import LocalWorkspace
+from .memory_workspace import MemoryWorkspace, default_workspace
